Move CLI/TOML parsing to private module
There is a lot of tricky code there that I marked as private
with underscores, but I think the entire module should be marked
private to avoid third-party code depending on it. In other words,
the parsing code is so specific to taps.run that no one else
should use it and there will very likely be internally breaking
changes in the future to support new features.
